refactor: extract shared modules under src/common/module
