+++ src/common/__init__.py
@@ -1,0 +1,1 @@
+"""Shared common modules for reusable pipeline components."""
