+++ src/common/module/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable common modules used across pipelines and services."""
