+++ src/pipeline/scraper/module/jepx.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 from dataclasses import dataclass
 from datetime import datetime
 
+from common.module.http_scraper import BaseHttpScraper, RequestSpec
 from pipeline.jepx.common import (
     resolve_fiscal_year,
     resolve_spot_summary_file_name,
     resolve_spot_summary_object_key,
 )
-from pipeline.scraper.module.base import BaseHttpScraper, RequestSpec
 
 
 @dataclass(frozen=True)
